+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### next
+
+[NEW] commands typing
+[NEW] commands support TS files
+[UPDATE] update deps
+
 ### 5.0.0-beta.9
 
 [BREAKING] move email module to separate package @adaptivestone/framework-module-email. Please use it if you want to send emails